+++ inst/doc/b-dimensions.R
@@ -3,5 +3,5 @@ library(rgho)
 library(knitr)
 
 ## ---- echo=FALSE--------------------------------------------------------------
-kable(rgho:::to_data_frame(get_gho_dimensions()))
+kable(get_gho_dimensions())