+++ inst/doc/e-indicators.R
@@ -3,5 +3,7 @@ library(rgho)
 library(knitr)
 
 ## ---- echo=FALSE--------------------------------------------------------------
-kable(rgho:::to_data_frame(get_gho_codes(dimension = "GHO")))
+kable(
+      get_gho_indicators()
+  )