+++ NEWS.md
@@ -1,6 +1,21 @@
+# rgho 3.0.0
+This new version of `rgho` now uses the most recent API of the WHO, 
+which simplifies considerably the package, and relies on the package 
+[ODataQuery.](https://CRAN.R-project.org/package=ODataQuery)
+
+## Breaking changes
+ * Objects created with get_gho_* are now of class "gho" and "data.frame"
+ * Former lower-level access to the API do not work anymore
+ * If a connection error occurs, the function returns an empty list with a 
+  message as argument 
+ * The proxy settings and the number of retries are now managed by `ODataQuery` 
+ package.
+ * There is no memoisation anymore
+ * `get_gho_codes()` is superseded by `get_gho_values()` to better reflect the 
+ new API terminology
+
 # rgho 2.0.2
   * better handles connection fails
-
 
 # rgho 2.0.1
   * fix errors caused by dplyr 0.8